Revert "Admin pages"
