+++ client/src/components/Signup/signUp.css
@@ -96,8 +96,3 @@
 	color: white;
 	margin: 10px;
 }
-
-.headerBlack {
-	color: rgb(220, 12, 12);
-}
-
