--- client/src/components/button/button.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const button = ({ type = "button", className = "", onClick, children }) => {
-  return (
-    <button type={type} className={className} onClick={onClick}>
-			{children}
-		</button>
-  )
-}
-
-export default button
