--- client/src/screens/About/About.jsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Layout/Header";
-import Footer from "../../components/Layout/Footer";
-
-const About = () => {
-
-  return (
-    <div className="">
-      <Header />
-      This is About page
-      <Footer />
-    </div>
-  );
-};
-
-export default About;
