--- client/src/screens/Admin/Billing/Billing.jsx
@@ -1,10 +1,0 @@
-const Billing = () => {
-
-  return (
-    <div className="">
-      This is Billing page
-    </div>
-  );
-};
-
-export default Billing;
