--- client/src/screens/Admin/Dashboard/Dashboard.jsx
@@ -1,10 +1,0 @@
-const Dashboard = () => {
-
-  return (
-    <div className="">
-      This is Dashboard page
-    </div>
-  );
-};
-
-export default Dashboard;
