--- client/src/screens/Admin/Invoice/Invoice.jsx
@@ -1,10 +1,0 @@
-const Invoice = () => {
-
-  return (
-    <div className="">
-      This is Invoice page
-    </div>
-  );
-};
-
-export default Invoice;
