--- client/src/screens/Admin/Order/Order.jsx
@@ -1,10 +1,0 @@
-const Order = () => {
-
-  return (
-    <div className="">
-      This is Order page
-    </div>
-  );
-};
-
-export default Order;
