--- client/src/screens/Admin/Product/Product.jsx
@@ -1,10 +1,0 @@
-const Product = () => {
-
-  return (
-    <div className="">
-      This is Product page
-    </div>
-  );
-};
-
-export default Product;
