--- client/src/screens/Admin/User/User.jsx
@@ -1,10 +1,0 @@
-const User = () => {
-
-  return (
-    <div className="">
-      This is User page
-    </div>
-  );
-};
-
-export default User;
