--- client/src/screens/Blog/Blog.jsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Layout/Header";
-import Footer from "../../components/Layout/Footer";
-
-const Blog = () => {
-
-  return (
-    <div className="">
-      <Header />
-      This is Blog page
-      <Footer />
-    </div>
-  );
-};
-
-export default Blog;
