--- client/src/screens/Contact/Contact.jsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Layout/Header";
-import Footer from "../../components/Layout/Footer";
-
-const Contact = () => {
-
-  return (
-    <div className="">
-      <Header />
-      This is Contact page
-      <Footer />
-    </div>
-  );
-};
-
-export default Contact;
