--- client/src/screens/Features/Feature.jsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Layout/Header";
-import Footer from "../../components/Layout/Footer";
-
-const Feature = () => {
-
-  return (
-    <div className="">
-      <Header />
-      This is feature page
-      <Footer />
-    </div>
-  );
-};
-
-export default Feature;
