--- client/src/screens/Home/HomePage.jsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Layout/Header";
-import Footer from "../../components/Layout/Footer";
-
-const HomePage = () => {
-
-  return (
-    <div className="">
-      <Header />
-      This is home page
-      <Footer />
-    </div>
-  );
-};
-
-export default HomePage;
